+++ wp-content/plugins/aelia-currencyswitcher-subscriptions/src/lib/classes/integration/woothemes-subscriptions-plugin/wc-aelia-cs-subscriptions-integration.php
@@ -239,7 +239,8 @@ class Subscriptions_Integration {
 		//	$product->subscription_sign_up_fee
 		//);
 
-		if(is_numeric($product->sale_price) && $product->is_on_sale()) {
+		if(is_numeric($product->sale_price) &&
+			 $this->product_is_on_sale($product, $product->sale_price, $product->regular_price)) {
 			$product->price = $product->sale_price;
 		}
 		else {
@@ -410,15 +411,20 @@ class Subscriptions_Integration {
 		// Fix checkout currency during renewals
 		// NOTE: as of 15/12/2015 this patch should no longer be needed.
 		//add_filter('woocommerce_order_again_cart_item_data', array($this, 'woocommerce_order_again_cart_item_data'), 10, 3);
-		//add_filter('woocommerce_checkout_init', array($this, 'maybe_override_currency'), 10);
+		//add_filter('woocommerce_checkout_init', array($this, 'maybe_override_currency'), 50);
 
 		// Handle manual creation of subscription
 		// @since 1.3.8.171004
 		add_action('woocommerce_process_shop_order_meta', array($this, 'woocommerce_process_shop_order_meta'), 5, 2);
 		// Currency selector on the Edit Subscription page
 		// @since 1.3.8.171004
-		add_action('woocommerce_currency', array($this, 'get_currency_for_manual_subscription'), 35, 1);
+		add_filter('woocommerce_currency', array($this, 'get_currency_for_manual_subscription'), 35, 1);
 		add_action('add_meta_boxes', array($this, 'add_meta_boxes'));
+
+		// Load the Edit Order scripts on the Edit Subscription page, to ensure that
+		// product prices are loaded in the correct order currency
+		// @since 1.3.11.180222
+		add_filter('wc_aelia_cs_load_order_edit_scripts', array($this, 'wc_aelia_cs_load_order_edit_scripts'), 10, 2);
 	}
 
 	/**
@@ -656,7 +662,14 @@ class Subscriptions_Integration {
 			if($this->product_requires_conversion($product, $selected_currency)) {
 				$product = $this->convert_product_prices($product, $selected_currency);
 			}
-			$subscription_sign_up_fee = $product->subscription_sign_up_fee;
+
+			// Check that the sign up fee is numeric, before returning it. This is to
+			// prevent warnings being raised by the Subscriptions plugin, which takes
+			// fees "blindly"
+			// @since 1.3.10.180218
+			if(is_numeric($product->subscription_sign_up_fee)) {
+				$subscription_sign_up_fee =  (float)$product->subscription_sign_up_fee;
+			}
 		}
 
 		return $subscription_sign_up_fee;
@@ -900,13 +913,18 @@ class Subscriptions_Integration {
 	 * @since 1.3.8.171004
 	 */
 	public function woocommerce_process_shop_order_meta($post_id, $post) {
+		if($post->post_type != 'shop_subscription') {
+			return;
+		}
+
 		// Set the currency on manually created orders when their first draft is saved.
 		// This is done to prevent WooCommerce from returning shop's base currency
 		// when WC_Order::get_currency() is called. See old code below, for reference
 		// @since 4.5.1.171012
 		$order = wc_get_order($post_id);
-		if(in_array($order->get_status(), array('draft', 'auto-draft'))) {
-			add_filter('woocommerce_currency', array($this->currency_switcher(), 'woocommerce_currency'), 5);
+
+		if($order->has_status(array('draft', 'auto-draft', 'pending')) && !empty($_POST['aelia_cs_currency'])) {
+			add_filter('woocommerce_currency', array($this->currency_switcher(), 'woocommerce_currency'), 50);
 		}
 
 		// Only set the currency if the order doesn't have one set against it.
@@ -1044,5 +1062,73 @@ class Subscriptions_Integration {
 	protected function displayed_order_currency() {
 		global $post;
 		return $this->currency_switcher()->get_order_currency($post->ID);
+	}
+
+	/**
+	 * Ensure that the JavaScript for the Edit Order page are also loaded on the
+	 * Add/Edit Subscription page.
+	 *
+	 * @param bool should_load_scripts
+	 * @param object post
+	 * @return bool
+	 * @since 1.3.11.180222
+	 */
+	public function wc_aelia_cs_load_order_edit_scripts($should_load_scripts, $post) {
+		if(!$should_load_scripts) {
+			$post_type = is_object($post) ? $post->post_type : null;
+			$should_load_scripts = ($post_type === 'shop_subscription');
+		}
+
+		return $should_load_scripts;
+	}
+
+	/**
+	 * Indicates if the product is on sale. A product is considered on sale if:
+	 * - Its "sale end date" is empty, or later than today.
+	 * - Its sale price in the active currency is lower than its regular price.
+	 *
+	 * @param WC_Product product The product to check.
+	 * @param mixed sale_price The product's sale price. If null, the sale price
+	 * is fetched by calling WC_Product::get_sale_price().
+	 * @param mixed regular_price The product's regular price. If null, the regular price
+	 * is fetched by calling WC_Product::get_regular_price().
+	 * @return bool
+	 * @since 1.3.12.180713
+	 */
+	protected function product_is_on_sale(WC_Product $product, $sale_price = null, $regular_price = null) {
+		$sale_price_dates_from = $this->date_to_string($product->get_date_on_sale_from());
+		$sale_price_dates_to = $this->date_to_string($product->get_date_on_sale_to());
+
+		$is_on_sale = false;
+		$today = date('Ymd');
+		if((empty($sale_price_dates_from) ||
+				$today >= $sale_price_dates_from) &&
+			 (empty($sale_price_dates_to) ||
+				$sale_price_dates_to > $today)) {
+			$sale_price = $sale_price !== null ? $sale_price : $product->get_sale_price();
+			$regular_price = $regular_price !== null ? $regular_price : $product->get_regular_price();
+
+			$is_on_sale = is_numeric($sale_price) && ($sale_price < $regular_price);
+		}
+		return $is_on_sale;
+	}
+
+	/**
+	 * Converts a timestamp, or a date object, to the specified format.
+	 *
+	 * @param int|WC_Datetime date The date to convert.
+	 * @param string format The target format.
+	 * @return string The date as a string in YMD format.
+	 * @since 1.3.12.180713
+	 */
+	protected function date_to_string($date, $format = 'Ymd') {
+		if(empty($date)) {
+			return '';
+		}
+
+		if(is_object($date) && ($date instanceof \WC_DateTime)) {
+			return $date->format($format);
+		}
+		return date($format, $date);
 	}
 }